refine create function

+++ src/client.c
@@ -39,7 +39,7 @@ machine please look into this as a a source of error. */
  **/
 int connect_client() {
     int client_socket;
-    size_t len;
+    int len;
     struct sockaddr_un remote;
 
     log_info("Attempting to connect...\n");
